back/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
pymysql>=1.1.0
cryptography>=42.0.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
pydantic[email]>=2.6.0
alembic>=1.13.0
pytest>=8.0.0
httpx>=0.26.00
pytest-asyncio>=0.23.0
python-dotenv>=1.0.0
